+++ audio_engine/ReverbController.h
@@ -23,11 +23,11 @@ namespace CloudSeed
 
 		ReverbChannel channelL;
 		ReverbChannel channelR;
-		double leftChannelIn[bufferSize];
-		double rightChannelIn[bufferSize];
-		double leftLineBuffer[bufferSize];
-		double rightLineBuffer[bufferSize];
-		double parameters[(int)Parameter::Count];
+		float leftChannelIn[bufferSize];
+		float rightChannelIn[bufferSize];
+		float leftLineBuffer[bufferSize];
+		float rightLineBuffer[bufferSize];
+		float parameters[(int)Parameter::Count];
 
 	public:
 		ReverbController(int samplerate)
@@ -55,12 +55,12 @@ namespace CloudSeed
 			return (int)Parameter::Count;
 		}
 
-		double* GetAllParameters()
+		float* GetAllParameters()
 		{
 			return parameters;
 		}
 
-		double GetScaledParameter(Parameter param)
+		float GetScaledParameter(Parameter param)
 		{
 			switch (param)
 			{
@@ -138,7 +138,7 @@ namespace CloudSeed
 			return 0.0;
 		}
 
-		void SetParameter(Parameter param, double value)
+		void SetParameter(Parameter param, float value)
 		{
 			parameters[(int)param] = value;
 			auto scaled = GetScaledParameter(param);
@@ -153,7 +153,7 @@ namespace CloudSeed
 			channelR.ClearBuffers();
 		}
 
-		void Process( const float** input, float** output, int bufferSize)
+		void Process(const float** input, float** output, int bufferSize)
 		{
 			auto len = bufferSize;
 			auto cm = GetScaledParameter(Parameter::InputMix) * 0.5;
@@ -178,7 +178,7 @@ namespace CloudSeed
 		}
 		
 	private:
-		double P(Parameter para)
+		float P(Parameter para)
 		{
 			auto idx = (int)para;
 			return idx >= 0 && idx < (int)Parameter::Count ? parameters[idx] : 0.0;
